优化以及修改，所有接口自测通过

+++ src/main/java/com/school/manager/service/impl/ChapterServiceImpl.java
@@ -1,20 +1,25 @@
 package com.school.manager.service.impl;
 
 import com.school.manager.common.constant.LongConstant;
+import com.school.manager.enums.StatusCode;
+import com.school.manager.exception.SysServiceException;
+import com.school.manager.jwt.LoginUserInfo;
+import com.school.manager.jwt.LoginUserUtil;
 import com.school.manager.pojo.dao.ChapterDao;
 import com.school.manager.pojo.dto.req.ChapterSaveReq;
 import com.school.manager.pojo.dto.req.ChapterUpdateReq;
 import com.school.manager.pojo.dto.resp.ChapterInfoResp;
-import com.school.manager.enums.StatusCode;
-import com.school.manager.exception.SysServiceException;
+import com.school.manager.pojo.dto.resp.GradeSubjectResp;
 import com.school.manager.pojo.entity.Chapter;
 import com.school.manager.service.ChapterService;
 import com.school.manager.utils.BeanMapper;
 import com.school.manager.utils.IdWorker;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
@@ -37,8 +42,12 @@ public class ChapterServiceImpl implements ChapterService {
      */
     @Override
     public String save(ChapterSaveReq request) {
+        // 转换对象
         Chapter chapter = BeanMapper.def().map(request, Chapter.class);
         chapter.setId(String.valueOf(idWorker.nextId()));
+        // 若为当前节点顶级节点，则rootId设为当前节点id
+        String rootId = StringUtils.equals("-1", chapter.getParentId()) ? chapter.getId() : Optional.ofNullable(chapterDao.info(chapter.getParentId())).map(Chapter::getRootId).orElseThrow(() -> new SysServiceException(StatusCode.PARENT_NOT_EXIST.getDesc()));
+        chapter.setRootId(rootId);
         chapter.setVersion(LongConstant.ONE);
         int insert = chapterDao.save(chapter);
         if (insert == 0) {
@@ -54,7 +63,9 @@ public class ChapterServiceImpl implements ChapterService {
      */
     @Override
     public void delete(String id) {
-        Chapter source = Optional.ofNullable(chapterDao.info(id)).orElseThrow(() -> new SysServiceException(StatusCode.DATA_NOT_EXIST.getDesc()));
+        if (chapterDao.childNodes(id) > 0) {
+            throw new SysServiceException(StatusCode.EXIST_CHILD_NODE.getDesc());
+        }
         chapterDao.delete(id);
     }
 
@@ -72,7 +83,9 @@ public class ChapterServiceImpl implements ChapterService {
             throw new SysServiceException(StatusCode.DATA_CHANGED.getDesc());
         }
         Chapter chapter = BeanMapper.def().map(request, Chapter.class);
+        String rootId = StringUtils.equals("-1", chapter.getParentId()) ? chapter.getId() : Optional.ofNullable(chapterDao.info(chapter.getParentId())).map(Chapter::getRootId).orElseThrow(() -> new SysServiceException(StatusCode.PARENT_NOT_EXIST.getDesc()));
         chapter.setVersion(chapter.getVersion() + LongConstant.ONE);
+        chapter.setRootId(rootId);
         chapterDao.update(chapter);
         return chapter.getId();
     }
@@ -87,5 +100,16 @@ public class ChapterServiceImpl implements ChapterService {
     public ChapterInfoResp info(String id) {
         Chapter chapter = Optional.ofNullable(chapterDao.info(id)).orElseThrow(() -> new SysServiceException(StatusCode.DATA_NOT_EXIST.getDesc()));
         return BeanMapper.def().map(chapter, ChapterInfoResp.class);
+    }
+
+    /**
+     * 当前用户所绑定的年级学科下拉列表
+     *
+     * @return 当前用户所绑定的年级学科下拉列表
+     */
+    @Override
+    public List<GradeSubjectResp> gradeSubjectDropdownList() {
+        String userId = Optional.ofNullable(LoginUserUtil.getLoginUserInfo()).map(LoginUserInfo::getId).orElseThrow(() -> new SysServiceException(StatusCode.NO_LOGIN_INFO.getDesc()));
+        return chapterDao.gradeSubjectDropdownList(userId);
     }
 }
